Ultimas rotas

+++ src/routes/diet_mels-routes.ts
@@ -9,66 +9,7 @@ import { z } from 'zod'
 import { checkSessionIdExists } from '../middlewares/check-session-id-exist'
 
 export async function dietMelsRoutes(app: FastifyInstance) {
-    
-    // TODO: LISTA UMA ÚNICA TRANSAÇÃO
-    // app.get('/:id', { preHandler: [checkSessionIdExist] }, async (request) => {
-    //     const getMelsParamsSchema = z.object({
-    //         id: z.string().uuid(),
-    //     })
-    //     const {id} = getMelsParamsSchema.parse(request.params)
-    //     const {sessionId} = request.cookies
-    //     const transaction = await knex('record_meal')
-    //     .where({
-    //         id,
-    //         session_id: sessionId
-    //     })
-    //     .first()
-    //     return {
-    //         transaction
-    //     }
-    // })
-    
-    // TODO: SOMA OS VALORES DE AMOUNT
-    // app.get('/summary', { preHandler: [checkSessionIdExist] }, async (request) => {
-    //     const {sessionId} = request.cookies
-    //     const summary = await knex('transactions')
-    //     .where('session_id', sessionId)
-    //     .sum('amount', {as: 'amount'})
-    //     .first()
-    //     return summary
-    // })
-    
-    // TODO: CRIA UMA NOVA TRANSAÇÃO
-    // app.post('/', async (request, reply) => {
-    //     const createDietBodySchema = z.object({
-    //         first_name: z.string(),
-    //         last_name: z.string(),
-    //         email: z.string(),
-    //     })
-    //     const { first_name, last_name, email } = createDietBodySchema.parse(request.body)
-        
-    //     // Valida o usuário pelo id no cookies
-    //     let sessionId = request.cookies.sessionId
-
-    //     // Se nao tiver o id do cookie, vai criar e será valido por 7 dias.
-    //     if(!sessionId){
-    //         sessionId = crypto.randomUUID()
-    //         reply.cookie('sessionId', sessionId, {
-    //             path: '/', // ESSE PATH ESPECIFICA QUAIS ROTAS IRÃO PODER ACESSAR ESTE COOKIE
-    //             maxAge: 60 * 60 * 24 * 7 // 7 DIAS
-    //         })
-    //     }
-    //     await knex('user_diet').insert({
-    //         id: crypto.randomUUID(),
-    //         first_name,
-    //         last_name,
-    //         email,
-    //         session_id: sessionId
-    //     })
-    //     return reply.status(201).send()
-    // })
     app.post('/', { preHandler: [checkSessionIdExists] }, async (request, reply) => {
-
         const createDietBodySchema = z.object({
             title: z.string(),
             description: z.string(),
@@ -77,7 +18,6 @@ export async function dietMelsRoutes(app: FastifyInstance) {
         })
 
         const {title, description, isOnDiet, date} = createDietBodySchema.parse(request.body)
-
         await knex('meals')
         .insert({
             id: crypto.randomUUID(),
@@ -90,12 +30,29 @@ export async function dietMelsRoutes(app: FastifyInstance) {
         return reply.status(201).send()
     })
 
-     // TODO: LISTA AS REFEICOES
+     // TODO: LISTA TODAS AS REFEICOES
     app.get('/', { preHandler: [checkSessionIdExists] }, async (request, reply) => {
         const meals= await knex('meals')
         .where({ users_id: request.users?.id })
         .orderBy('date', 'desc')   
         return reply.send({ meals })
+    })
+
+     // TODO: FILTRA UMA ÚNICA REFEIÇÃO ENTRE TODAS DE ACORDO COM O ID
+     app.get('/:id', { preHandler: [checkSessionIdExists] }, async (request) => {
+        const getTransactionsParamsSchema = z.object({
+            id: z.string().uuid(),
+        })
+        const {id} = getTransactionsParamsSchema.parse(request.params)
+        const uniqMels = await knex('meals')
+        .where({
+            id,
+            users_id: request.users?.id
+        })
+        .first()
+        return {
+            uniqMels
+        }
     })
 
     // TODO: ALTERA AS REFEIÇÕES
@@ -110,7 +67,6 @@ export async function dietMelsRoutes(app: FastifyInstance) {
         })
         const {id} = getMelsParamsSchema.parse(request.params)
         const { title, description, isOnDiet} = createDietBodySchema.parse(request.body)
-
         await knex('meals')
         .where({
             id,
@@ -128,10 +84,49 @@ export async function dietMelsRoutes(app: FastifyInstance) {
     app.delete('/:id', { preHandler: [checkSessionIdExists] }, async (request, reply) => {
         const getMelsParamsSchema = z.object({id: z.string().uuid()})
         const {id} = getMelsParamsSchema.parse(request.params)    
-
+        const validaMels = await knex('meals').where({id:id}).first()
+        if(!validaMels){
+            return reply.status(404).send({error:'Met not found'})
+        }
         await knex('meals')
         .where({id:id})
         .delete()
         return reply.status(201).send()
+    })
+
+    // TODO: CRIA OS CALCULOS DAS MÉTRICAS
+
+    app.get('/metrics', { preHandler: [checkSessionIdExists] }, async (request, reply) => {
+        const totalDietClean = await knex('meals')
+            .where({users_id: request.users?.id, is_on_diet: true})// Pegas os users_id e os is_on_diet que forem true
+            .count('id', {as: 'total'})
+            .first()
+        const totalDietDirty = await knex('meals')
+            .where({users_id: request.users?.id, is_on_diet: false})// Pegas os users_id e os is_on_diet que forem false
+            .count('id', {as: 'total'})
+            .first()
+        const totalMels = await knex('meals')
+            .where({users_id: request.users?.id})
+            .orderBy('date', 'desc')
+            
+        const { dietClean } = totalMels.reduce((acc, meal) => {
+            if(meal.is_on_diet){
+                acc.current += 1
+            }else{
+                acc.current = 0
+            }
+            if(acc.current > acc.dietClean){
+                acc.dietClean = acc.current
+            }
+            return acc
+        },
+        {dietClean: 0, current: 0},
+        )
+        return reply.send({
+            totalMels: totalMels.length,
+            totalDietClean: totalDietClean?.total,
+            totalDietDirty: totalDietDirty?.total,
+            dietClean
+        })
     })
 }
